added db.js file
