+++ spec.js
@@ -1,20 +1,5 @@
 const { expect } = require('chai');
-const Sequelize = require('sequelize');
-console.log(process.env.DATABASE_URL);
-const conn = new Sequelize(process.env.DATABASE_URL);
-const { STRING } = Sequelize;
-const syncAndSeed = async()=> {
-  await conn.sync({ force: true });
-  await Promise.all([
-    User.create({ name: 'moe'}),
-    User.create({ name: 'lucy'}),
-    User.create({ name: 'ethyl'}),
-  ]);
-};
-
-const User = conn.define('user', ()=> {
-  name: STRING
-});
+const { models: { User}, syncAndSeed } = require('./db');
 
 describe('Math', ()=> {
   describe('1 + 1', ()=> {
